Move DDI abstractions to ddi module
Given that I added the kstat abstractions under ::opte::ddi, I think
it makes sense to move the other illumos DDI abstractions under that
module path as well. This is one of those things that is subjective,
the shed would look good in many different shades of blue; but at the
moment I find it nice to have a place to centralize the abstractions
that map to the DDI. In the fullness of time, there would be a wholly
independent crate that bundles these abstractions, but until that day
comes a dedicated module in opte should suffice.

This is separate from some of the other DDI abstractions used by xde,
e.g. the MacClient. I'm not sure if we actually consider the mac
client APIs part of the DDI today, so there's that twist. But in the
future, we might want to take an intermediate step between now and the
future with the independent crate by moving all safe DDI/illumos
private API abstractions to a dedicated crate under the opte repo.
There used to exist an illumos-ddi-dki crate, but I realized that
crate was misnamed a bit and so I merged its contents into
illumos-sys-hdrs. I think something like a new illumos-ddi crate,
perhaps with a "private" feature that enables abstractions for the
private illumos APIs would be a good next step; but it's a step that
doesn't need to happen today. I just wanted to add a little more
organization than currently existed.
